feat!: Instrument.TickSize is now double (was decimal)
Aligns with consumers (notably Tickr) that run price math in double for
speed. TickSet builder's price-overload `Add(int, PriceKind, decimal, int)`
becomes `Add(int, PriceKind, double, int)` for the same reason. Public API
flips from decimal → double on TickSize; PointValue was already double.

- Instrument.TickSize: decimal → double. BuildCache literals drop the m-suffix.
- Instrument.Round(double): uses double TickSize directly (no decimal cast).
- Instrument.TicksPerPoint: derived via Math.Round(1.0 / tickSize).
- TickSet.Builder.Add(decimal price, ...) → Add(double price, ...).
- TickSet enumerator multiplies in double; the (double) cast on TickSize is gone.
- RenkoCandleSet still uses decimal *internally* so brick-boundary math stays
  exact after long chains. The double TickSize is cast once at construction.
- Tests' m-suffix literals stripped throughout (sed pass).

BREAKING: consumers using `decimal Instrument.TickSize` or the
`Add(int, PriceKind, decimal, int)` overload must move to double.

+++ tests/SquidEyes.Pricing.UnitTests/Instruments/InstrumentTests.cs
@@ -15,7 +15,7 @@ public class InstrumentTests
     [InlineData(Symbol.M2K, 0.10)]
     [InlineData(Symbol.MGC, 0.10)]
     [InlineData(Symbol.MCL, 0.01)]
-    public void Create_ReturnsCorrectTickSize(Symbol symbol, decimal expectedTickSize)
+    public void Create_ReturnsCorrectTickSize(Symbol symbol, double expectedTickSize)
     {
         var instrument = Instrument.Create(symbol);
         Assert.Equal(expectedTickSize, instrument.TickSize);
